Download functionality for the tool
I tried to add the download functionality to the tool, I'm not sure if it works, I couldn't test it because auto-refresh wasn't working on my client. I was expecting you could make something better out of it, I copied part of the E2 code because I'm not really a gmod scripter.